% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_version.R
\name{get_current_version}
\alias{get_current_version}
\title{Get Current - internal helper function}
\usage{
get_current_version(file, version_control, verbose = FALSE)
}
\arguments{
\item{file}{name of file to be read}

\item{version_control}{version control file}

\item{verbose}{logical, show record returned?}
}
\value{
character string
}
\description{
Get Current - internal helper function
}
\keyword{internal}
