% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vc_name.R
\name{get_vc_name}
\alias{get_vc_name}
\title{Copy the stamped file name to the clip board}
\usage{
get_vc_name(file, version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{file}{name of file to retrieve stamped name}

\item{version_control}{name of version control file, default is `DATA_VC`}

\item{verbose}{logical, print output to console?}
}
\value{
writes string to system clipboard, invisibly
}
\description{
Given a file, gets its stamped name from `DATA_VC` for cases when
you want to hardcode a file reference. The stamped file name is pushed to the
system clipboard using the `clipr` package and can be immediately pasted.
}
\examples{
\donttest{
make_data_vc()
saveRDS_vc(cars, "my_cars.rds", verbose = TRUE)
get_vc_name("my_cars.rds")
}
}
