% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_vc.R
\name{make_data_vc}
\alias{make_data_vc}
\title{Initialize version control `DATA_VC` file}
\usage{
make_data_vc(file = "DATA_VC", replace = FALSE)
}
\arguments{
\item{file}{default is "DATA_VC, for optimal use do not change.}

\item{replace}{Allows overwriting of data_vc file.}
}
\value{
null invisibly
}
\description{
Function should be run on project startup to initialize the
data version control file. For optimal use, do not change file name.
}
\examples{
make_data_vc()
}
