% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read.csv_vc}
\alias{read.csv_vc}
\title{Wrapper for read.csv}
\usage{
read.csv_vc(file, ..., version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{...}{other params for read.csv see \code{\link[utils]{read.csv}}}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report returned current file version? Default is FALSE}
}
\value{
a string `stamp`_`file`
}
\description{
This is a wrapper function for read.csv. You can use it exactly as you would read.csv.
Internally it checks the data version control file and returns the stamp associated
with the current version of the desired `file`.
}
\examples{
\donttest{
make_data_vc()
write.csv_vc(cars, "my_cars.csv", verbose = TRUE)
read.csv_vc("my_cars.csv", verbose = TRUE)
}
}
