% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{read.dta_vc}
\alias{read.dta_vc}
\title{Read Stata Files, using pkg foreign}
\usage{
read.dta_vc(file, ..., version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{file}{a filename or URL as a character string.}

\item{...}{other params for read.dta see \code{\link[foreign]{read.dta}}}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report returned current file version? Default is FALSE}
}
\value{
A data frame with attributes.
}
\description{
This is a wrapper function for `foreign::read.dta`, use is identical to the original function.
}
\details{
Internally it checks the data version control file and returns the stamp associated
with the current version of the desired `file`.
}
\examples{
\donttest{
make_data_vc()
foreign::write.dta_vc(cars, "my_cars.dta", verbose = TRUE)
foreign::read.dta_vc("my_cars.dta", verbose = TRUE)
}
}
\seealso{
\code{\link[foreign]{read.dta}}
}
