% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv_vc}
\alias{read_csv_vc}
\title{Wrapper for readr::read_csv}
\usage{
read_csv_vc(file, ..., version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{...}{other params for read.csv see \code{\link[readr]{read_csv}}}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report returned current file version? Default is FALSE}
}
\value{
a string `stamp`_`file`
}
\description{
This is a wrapper function for readr::read_csv. You can use it exactly as you would read_csv.
Internally it checks the data version control file and returns the stamp associated
with the current version of the desired `file`.
}
\examples{
\donttest{
make_data_vc()
write_csv_vc(cars, "my_cars.csv", verbose = TRUE)
read_csv_vc("my_cars.csv", verbose = TRUE)
}
}
