% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{readRDS_vc}
\alias{readRDS_vc}
\title{Wrapper for readRDS}
\usage{
readRDS_vc(file, ..., version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{file}{a connection or the name of the file where the `R` object is saved to or read from}

\item{...}{other params for readRDS see \code{\link[base]{readRDS}}}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report returned current file version? Default is FALSE}
}
\value{
a string `stamp`_`file`
}
\description{
This is a wrapper function for readRDS. You can use it exactly as you would readRDS.
Internally it checks the data version control file and returns the stamp associated
with the current version of the desired `file`.
}
\examples{
\donttest{
make_data_vc()
saveRDS_vc(cars, "my_cars.rds", verbose = TRUE)
readRDS_vc("my_cars.rds", verbose = TRUE)
}
}
