% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{saveRDS_vc}
\alias{saveRDS_vc}
\title{Wrapper for saveRDS}
\usage{
saveRDS_vc(object, file, ..., stamp = Sys.Date(),
  version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{object}{`R` object to serialize.}

\item{file}{a connection or the name of the file where the `R` object is saved to or read from}

\item{...}{other params for saveRDS see \code{\link[base]{saveRDS}}}

\item{stamp}{file suffix to disambiguate, default is Sys.Date()}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report update to version control file? Default is FALSE}
}
\value{
NULL invisibly
}
\description{
This is a wrapper for saveRDS, you can use it exactly as you would use saveRDS.
Internally it checks for the data version control file, and inserts or updates
the record for the file that you are saving. The tracking is applied to the `file`
}
\examples{
\donttest{
make_data_vc()
saveRDS_vc(cars, "my_cars.rds", verbose = TRUE)
}
}
\seealso{
\code{\link[base]{saveRDS}}
}
