% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_file.R
\name{stamp_file}
\alias{stamp_file}
\title{Add the stamp to the file name}
\usage{
stamp_file(file, stamp)
}
\arguments{
\item{file}{a text string with the user input file name}

\item{stamp}{the version control stamp to add}
}
\value{
character string
}
\description{
this is a utility function to add the stamp to the file name.
The stamp is added before the last period (if no periods are present it is
added at the end). The stamp is delinetaed with a leading and trailing
underscore (`_`).
}
\keyword{internal}
