% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version_control.r
\name{update_version_control}
\alias{update_version_control}
\title{Update Version Control}
\usage{
update_version_control(file, stamp, version_control, verbose = FALSE)
}
\arguments{
\item{file}{name of file being written}

\item{stamp}{value of stamp}

\item{version_control}{version control file}

\item{verbose}{logical, show record update values?}
}
\value{
invisibly updates
}
\description{
DATA_VC is a file that tracks which data files are currently
used in the program. This is used because git is not tracking changes in the data
files. Whenever a new file is saved or an old file updated,the record in DATA_VC
will be updated to include the current stamp. This function is used inside the
save wrapper functions (e.g. saveRDS_vc)
}
\keyword{internal}
