% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta.R
\name{write.dta_vc}
\alias{write.dta_vc}
\title{Write Stata Files, using pkg foreign}
\usage{
write.dta_vc(dataframe, file, ..., stamp = Sys.Date(),
  version_control = "DATA_VC", verbose = FALSE)
}
\arguments{
\item{dataframe}{a data frame}

\item{file}{character string giving filename.}

\item{...}{other params for write.dta see \code{\link[foreign]{write.dta}}}

\item{stamp}{file suffix to disambiguate, default is Sys.date()}

\item{version_control}{name of data version control file. Default is `data_vc`. Should match name used in `make_data_vc`.}

\item{verbose}{logical, report update to version control file? Default is FALSE}
}
\description{
This is a wrapper for `foreign::write.dta()`, use is identical to the original function.
}
\details{
Internally it checks for the data version control file, and inserts or updates
the record for the file that you are saving. The tracking is applied to the `file`
}
\examples{
\donttest{
make_data_vc()
foreign::write.dta_vc(cars, "my_cars.dta", verbose = TRUE)
}
}
\seealso{
\code{\link[foreign]{write.dta}}
}
